Move JavaCCTokenizer to internal package
