+++ pmd-javascript/src/main/java/net/sourceforge/pmd/cpd/EcmascriptTokenizer.java
@@ -6,6 +6,7 @@ package net.sourceforge.pmd.cpd;
 
 import java.io.StringReader;
 
+import net.sourceforge.pmd.cpd.internal.JavaCCTokenizer;
 import net.sourceforge.pmd.lang.TokenManager;
 import net.sourceforge.pmd.lang.ast.GenericToken;
 import net.sourceforge.pmd.lang.ecmascript5.Ecmascript5TokenManager;
