+++ pmd-matlab/src/main/java/net/sourceforge/pmd/cpd/MatlabTokenizer.java
@@ -6,6 +6,7 @@ package net.sourceforge.pmd.cpd;
 
 import java.io.StringReader;
 
+import net.sourceforge.pmd.cpd.internal.JavaCCTokenizer;
 import net.sourceforge.pmd.lang.TokenManager;
 import net.sourceforge.pmd.lang.matlab.MatlabTokenManager;
 import net.sourceforge.pmd.util.IOUtil;
