+++ pmd-plsql/src/main/java/net/sourceforge/pmd/cpd/PLSQLTokenizer.java
@@ -7,6 +7,7 @@ package net.sourceforge.pmd.cpd;
 import java.io.StringReader;
 import java.util.Properties;
 
+import net.sourceforge.pmd.cpd.internal.JavaCCTokenizer;
 import net.sourceforge.pmd.lang.TokenManager;
 import net.sourceforge.pmd.lang.ast.GenericToken;
 import net.sourceforge.pmd.lang.plsql.PLSQLTokenManager;
