+++ pmd-python/src/main/java/net/sourceforge/pmd/cpd/PythonTokenizer.java
@@ -6,6 +6,7 @@ package net.sourceforge.pmd.cpd;
 
 import java.io.StringReader;
 
+import net.sourceforge.pmd.cpd.internal.JavaCCTokenizer;
 import net.sourceforge.pmd.lang.TokenManager;
 import net.sourceforge.pmd.lang.python.PythonTokenManager;
 import net.sourceforge.pmd.util.IOUtil;
